fix(HealthSquad): ajustes no Scroll da página
